requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
et-xmlfile==1.0.1
idna==2.8
idna-ssl==1.1.0
jdcal==1.4.1
kiwisolver==1.1.0
matplotlib==3.1.1
multidict==4.5.2
numpy==1.16.4
oauthlib==3.0.2
openpyxl==2.6.2
pandas==0.22.0
py==1.8.0
pyparsing==2.4.0
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
retry==0.9.2
six==1.12.0
slackclient==2.1.0
tweepy==3.8.0
typing-extensions==3.7.4
urllib3==1.25.3
yarl==1.3.0
